Changed launch time code to work

+++ SoundByte/NSMutableArray+Shuffling.m
@@ -16,7 +16,8 @@
     for (NSUInteger i = 0; i < count; ++i) {
         // Select a random element between i and end of array to swap with.
         NSInteger nElements = count - i;
-        NSInteger n = arc4random_uniform(nElements) + i;
+        //Suppressed Warning by signing the number of elements into a 32-bit architecture, not good really, but when will we have an array bigger than 2^31?
+        NSInteger n = arc4random_uniform((uint32_t)nElements) + i;
         [self exchangeObjectAtIndex:i withObjectAtIndex:n];
     }
 }
